impl ItsAProcess for Validator
